deployment/requirements.txt
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.4.3
matplotlib-inline==0.1.2
opencv-python==4.5.3.56
pandas==1.3.2
Pillow==8.3.2
scikit-learn==0.24.2
seaborn==0.11.2
sklearn==0.0
streamlit==0.88.0
tensorflow==2.6.0
